Completed Issue 277: Stockpile: Decimals

+++ src/main/java/net/nikr/eve/jeveasset/gui/shared/menu/JMenuStockpile.java
@@ -44,7 +44,7 @@ public class JMenuStockpile<T> extends JAutoMenu<T> {
 		ADD_TO
 	}
 
-	private static final int DEFAULT_ADD_COUNT = 1;
+	private static final double DEFAULT_ADD_COUNT = 1;
 
 	private List<Stockpile> stockpilesCashe = null;
 	private final List<JMenuItem> jMenuItems = new ArrayList<JMenuItem>();
